migrations/versions/68788e71959d_initial_migration.py
"""Initial migration

Revision ID: 68788e71959d
Revises: 
Create Date: 2026-01-04 21:38:25.341030

"""
from alembic import op
import sqlalchemy as sa


revision = '68788e71959d'
down_revision = None
branch_labels = None
depends_on = None


def upgrade():
    
    op.create_table('heroes',
    sa.Column('id', sa.Integer(), nullable=False),
    sa.Column('name', sa.String(length=100), nullable=False),
    sa.Column('super_name', sa.String(length=100), nullable=False),
    sa.PrimaryKeyConstraint('id')
    )
    op.create_table('powers',
    sa.Column('id', sa.Integer(), nullable=False),
    sa.Column('name', sa.String(length=100), nullable=False),
    sa.Column('description', sa.Text(), nullable=False),
    sa.PrimaryKeyConstraint('id')
    )
    op.create_table('hero_powers',
    sa.Column('id', sa.Integer(), nullable=False),
    sa.Column('strength', sa.String(length=20), nullable=False),
    sa.Column('hero_id', sa.Integer(), nullable=False),
    sa.Column('power_id', sa.Integer(), nullable=False),
    sa.ForeignKeyConstraint(['hero_id'], ['heroes.id'], ),
    sa.ForeignKeyConstraint(['power_id'], ['powers.id'], ),
    sa.PrimaryKeyConstraint('id')
    )
    


def downgrade():
    op.drop_table('hero_powers')
    op.drop_table('powers')
    op.drop_table('heroes')
    
